refactor and move SelfSignedCertificateGenerator into proxy-node-test src
